payment/requirements.txt
fastapi==0.75.0
redis-om==0.0.20
python-dotenv~=0.19.2
starlette~=0.17.1